+++ src/main/java/com/codeit/sb13/monew/article/service/impl/ArticleServiceImpl.java
@@ -1,6 +1,6 @@
 package com.codeit.sb13.monew.article.service.impl;
 
-import com.codeit.sb13.monew.article.controller.dto.ArticleDto;
+import com.codeit.sb13.monew.article.service.dto.ArticleDto;
 import com.codeit.sb13.monew.article.domain.Article;
 import com.codeit.sb13.monew.article.domain.ArticleSource;
 import com.codeit.sb13.monew.article.mapper.ArticleMapper;
@@ -47,7 +47,7 @@ public class ArticleServiceImpl implements ArticleService {
         long viewCount = articleViewRepository.countByArticle_Id(articleId);
 
         // commentCount는 댓글 파트 집계 방식 확정 전까지 0 (MID4-147)
-        return articleMapper.toDto(article, viewedByMe, 0, (int) viewCount);
+        return articleMapper.toDto(article, viewedByMe, 0L, viewCount);
     }
 
     @Override
